+++ tests/test_util.py
@@ -4,6 +4,7 @@ from argparse import Namespace
 from numpy.testing import assert_equal, assert_allclose
 import numpy as np
 from iminuit._core import MnUserParameterState
+from iminuit._optional_dependencies import optional_module_for
 import pickle
 
 
@@ -122,10 +123,10 @@ def test_FixedView_comparison_with_broadcasting():
     assert_equal(f, [False, False, False])
 
     # broadcasting
-    assert f == False
+    assert f == False  # noqa
     f[0] = True
     assert_equal(f, [True, False, False])
-    assert f != False
+    assert f != False  # noqa
 
 
 def test_Matrix():
@@ -205,11 +206,11 @@ def test_Param():
     assert p.value == 1.2
     assert p.error == 3.4
     assert p.merror is None
-    assert p.is_const == False
-    assert p.is_fixed == False
-    assert p.has_limits == True
-    assert p.has_lower_limit is True
-    assert p.has_upper_limit is False
+    assert not p.is_const
+    assert not p.is_fixed
+    assert p.has_limits
+    assert p.has_lower_limit
+    assert not p.has_upper_limit
     assert p.lower_limit == 42
     assert p.upper_limit is None
 
@@ -364,7 +365,7 @@ def test_FMin(errordef):
     assert fmin.algorithm == "foo"
     assert fmin.edm == 1.23456e-10
     assert fmin.edm_goal == 0.1
-    assert fmin.has_parameters_at_limit == False
+    assert not fmin.has_parameters_at_limit
     assert fmin.time == 1.2
 
     assert fmin == util.FMin(fm, "foo", 1, 2, 1, 0.1, 1.2)
@@ -784,3 +785,28 @@ def test_smart_sampling_1(fn_expected):
 def test_smart_sampling_2():
     with pytest.warns(RuntimeWarning):
         util._smart_sampling(np.log, 1e-10, 1, tol=1e-10)
+
+
+def test_optional_module_for_1():
+    with optional_module_for("foo"):
+        import iminuit  # noqa
+
+
+def test_optional_module_for_2():
+    from iminuit.warnings import OptionalDependencyWarning
+
+    with pytest.warns(
+        OptionalDependencyWarning, match="foo requires optional package 'foobarbaz'"
+    ):
+        with optional_module_for("foo"):
+            import foobarbaz  # noqa
+
+
+def test_optional_module_for_3():
+    from iminuit.warnings import OptionalDependencyWarning
+
+    with pytest.warns(
+        OptionalDependencyWarning, match="foo requires optional package 'foo'"
+    ):
+        with optional_module_for("foo", replace={"foobarbaz": "foo"}):
+            import foobarbaz  # noqa